orchestrator/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
pydantic-settings>=2.6.0
arq>=0.26.0
httpx>=0.27.0
cryptography>=42.0.0
